#78 녹화 기능 추가 :: electron 빌드 시 리소스 path 에러 수정 및 path converting 코드 리팩토링
1. convert_path.json 을 추가하여 리소스 converting 코드 수정
